+++ routes/web.php
@@ -9,7 +9,7 @@ $router->get('/', function () use ($router) {
 
 $router->post('/register', function (Request $request) {
     return app()->call(
-        'App\Http\Controllers\RegisterUserByEmail\RegisterUserByEmailController@register',
+        'App\Http\Controllers\Register\RegisterController@register',
         ['request' => $request]
     );
 });
