+++ tests/app/Http/Controllers/Register/RegisterControllerTest.php
@@ -1,13 +1,17 @@
 <?php
 
-namespace Tests\app\Http\Controllers\RegisterUserByEmail;
+namespace Tests\app\Http\Controllers\Register;
 
 use App\Repositories\DB_Repositories;
 use Mockery;
 use Tests\TestCase;
 
-class RegisterUserByEmailControllerTest extends TestCase
+class RegisterControllerTest extends TestCase
 {
+    protected function setUp(): void
+    {
+        parent::setUp();
+    }
 
     /**
      * @test
@@ -17,9 +21,7 @@ class RegisterUserByEmailControllerTest extends TestCase
         $response = $this->call(
             'POST',
             '/register',
-            [],
-            [],
-            [],
+            [], [], [],
             ['CONTENT_TYPE' => 'application/json'],
             json_encode([])
         );
@@ -39,20 +41,20 @@ class RegisterUserByEmailControllerTest extends TestCase
         $response = $this->call(
             'POST',
             '/register',
-            [],
-            [],
-            [],
+            [], [], [],
             ['CONTENT_TYPE' => 'application/json'],
             json_encode(['email' => 'invalido'])
         );
 
         $this->assertEquals(400, $response->status());
         $this->assertJsonStringEqualsJsonString(
-            json_encode(['error' => 'The email must be a valid email address'], JSON_PRETTY_PRINT),
+            json_encode(
+                ['error' => 'The email must be a valid email address'],
+                JSON_PRETTY_PRINT
+            ),
             $response->getContent()
         );
     }
-
 
     /**
      * @test
@@ -60,6 +62,7 @@ class RegisterUserByEmailControllerTest extends TestCase
     public function gets200AndReturnsApiKeyWhenEmailIsValidAndUserDoesNotExist(): void
     {
         $email = 'test@example.com';
+
 
         $mockRepo = Mockery::mock(DB_Repositories::class);
 
@@ -72,10 +75,7 @@ class RegisterUserByEmailControllerTest extends TestCase
         $mockRepo
             ->shouldReceive('insertUser')
             ->once()
-            ->with(
-                $email,
-                Mockery::type('string')
-            )
+            ->with($email, Mockery::type('string'))
             ->andReturnTrue();
 
         $mockRepo
@@ -84,17 +84,17 @@ class RegisterUserByEmailControllerTest extends TestCase
 
         $this->app->instance(DB_Repositories::class, $mockRepo);
 
+
         $response = $this->call(
             'POST',
             '/register',
-            [],
-            [],
-            [],
+            [], [], [],
             ['CONTENT_TYPE' => 'application/json'],
             json_encode(['email' => $email])
         );
 
         $this->assertEquals(200, $response->status());
+
         $content = json_decode($response->getContent(), true);
         $this->assertArrayHasKey('api_key', $content);
         $this->assertNotEmpty($content['api_key']);
@@ -122,10 +122,7 @@ class RegisterUserByEmailControllerTest extends TestCase
         $mockRepo
             ->shouldReceive('updateApiKey')
             ->once()
-            ->with(
-                $email,
-                Mockery::type('string')
-            )
+            ->with($email, Mockery::type('string'))
             ->andReturnTrue();
 
         $mockRepo
@@ -134,17 +131,17 @@ class RegisterUserByEmailControllerTest extends TestCase
 
         $this->app->instance(DB_Repositories::class, $mockRepo);
 
+
         $response = $this->call(
             'POST',
             '/register',
-            [],
-            [],
-            [],
+            [], [], [],
             ['CONTENT_TYPE' => 'application/json'],
             json_encode(['email' => $email])
         );
 
         $this->assertEquals(200, $response->status());
+
         $content = json_decode($response->getContent(), true);
         $this->assertArrayHasKey('api_key', $content);
         $this->assertNotEmpty($content['api_key']);
